Use checkpoints when fetching PRs
